Prototype
Before I excecuted my plan to create the buttons and category information on page load, I decided to create a simple prototype to test the "hidden" functions and make a caegory show, dissapear and change the view of different elements. This allowed me to focus strictly on how I would be able to implement my function of creating the elements automatically instead of manually, and that means that if you add more information to the ressurser.js script, they will display on the page.